+++ CHANGES.rst
@@ -6,6 +6,7 @@ Improvements:
  * MXKDeviceView: Make clear that device names are publicly readable (vector-im/riot-ios/issues/2662).
  * Privacy: Remove the bind true flag from 3PID adds in settings (vector-im/riot-ios/issues/2650).
  * Privacy: Remove the ability to set an IS at login/registration (vector-im/riot-ios/issues/2661).
+ * Privacy: Use wellknown to discover the IS of a custom HS (vector-im/riot-ios/issues/2686).
 
 Changes in MatrixKit in 0.10.2 (2019-08-08)
 ==========================================