+++ CHANGES.rst
@@ -1,3 +1,12 @@
+Changes in MatrixKit in 0.10.3 (2019-xx-xx)
+==========================================
+
+Improvements:
+ *
+
+Bug fix:
+ * Display correctly the revoked third-party invite.
+ 
 Changes in MatrixKit in 0.10.2 (2019-08-08)
 ==========================================